initial implementation
Accepts a path to a directory and returns a tar stream that is
suitable for passing in to a GZip stream to make an npm pack style
tarball.

Alternatively, the stream could be piped to a tar.Parse() transform
stream to get the files back out again, or passed to docker for use
as an ADD/COPY workalike.