use package/ as tar stream root path
This is a much closer match to what 'npm pack' generates.

Note that this small change quietly replaces the entire underlying tar
implementation with tar-fs/tar-stream, but keeps the tar package around
as a dev dependency so we run our tests against a different tar
implementation. This is intentional as a first line defence against
breaking tar compatibility.