fix bad permissions when used on Windows
The problem is that Windows doesn't have the type of permissions bits
that tar uses, and tar-fs doesn't provide a default since it expects to
enforce a default on the unpacking side. The tarballs produced by
strong-pack are likely to be unpacked by npm, which assumes proper
permissions are already set.

This should fix strongloop/strong-pm#317